+++ source/state/state_definitions.hpp
@@ -73,9 +73,9 @@ namespace zlState {
     public:
         auto static constexpr ID = "window_w";
         auto static constexpr name = "NA";
-        inline static constexpr int minV = 140.f;
-        inline static constexpr int maxV = 2800.f;
-        inline static constexpr int defaultV = 280.f;
+        inline static constexpr float minV = 140.f;
+        inline static constexpr float maxV = 2800.f;
+        inline static constexpr float defaultV = 280.f;
         inline auto static const range =
                 juce::NormalisableRange<float>(minV, maxV, 1.f);
     };
@@ -84,9 +84,9 @@ namespace zlState {
     public:
         auto static constexpr ID = "window_h";
         auto static constexpr name = "NA";
-        inline static constexpr int minV = 95.f;
-        inline static constexpr int maxV = 1900.f;
-        inline static constexpr int defaultV = 190.f;
+        inline static constexpr float minV = 95.f;
+        inline static constexpr float maxV = 1900.f;
+        inline static constexpr float defaultV = 190.f;
         inline auto static const range =
                 juce::NormalisableRange<float>(minV, maxV, 1.f);
     };
